依旧是一批大规模去mcr化

+++ src/main/java/com/apocalypse/caerulaarbor/block/HighmoreSpawnblockBlock.java
@@ -29,7 +29,7 @@ import java.util.List;
 
 public class HighmoreSpawnblockBlock extends BaseEntityBlock implements SimpleWaterloggedBlock, EntityBlock {
 	public static final IntegerProperty BLOCKSTATE = IntegerProperty.create("blockstate", 0, 1);
-	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, (int) 2);
+	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, 2);
 	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
 	public HighmoreSpawnblockBlock() {
