+++ src/main/java/com/apocalypse/caerulaarbor/block/HighmoreSpawningBlockBlock.java
@@ -47,7 +47,7 @@ import java.util.List;
 
 public class HighmoreSpawningBlockBlock extends BaseEntityBlock implements SimpleWaterloggedBlock, EntityBlock {
 	public static final IntegerProperty BLOCKSTATE = IntegerProperty.create("blockstate", 0, 1);
-	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, (int) 2);
+	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, 2);
 	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
 	public HighmoreSpawningBlockBlock() {
