+++ src/main/java/com/apocalypse/caerulaarbor/block/RedOvaryBlock.java
@@ -3,7 +3,6 @@ package com.apocalypse.caerulaarbor.block;
 import com.apocalypse.caerulaarbor.configuration.CaerulaConfigsConfiguration;
 import com.apocalypse.caerulaarbor.init.CaerulaArborModGameRules;
 import com.apocalypse.caerulaarbor.network.CaerulaArborModVariables;
-import com.apocalypse.caerulaarbor.procedures.SummonRandomSeabornProcedure;
 import com.apocalypse.caerulaarbor.utils.EntityUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -67,7 +66,7 @@ public class RedOvaryBlock extends AbstractOvaryBlock {
 				}
 				chance = blockstate.getValue(OUTPUT);
 				if (Math.random() < chance * 0.005) {
-					if (EntityUtils.getSeabornNum(world, x, y, z) < Math.min((((LevelAccessor) world).getLevelData().getGameRules().getInt(CaerulaArborModGameRules.CLONE_NUMBER_LIMIT)), (double) CaerulaConfigsConfiguration.CLONE_NUM.get())) {
+					if (EntityUtils.getSeabornNum(world, x, y, z) < Math.min((((LevelAccessor) world).getLevelData().getGameRules().getInt(CaerulaArborModGameRules.CLONE_NUMBER_LIMIT)), CaerulaConfigsConfiguration.CLONE_NUM.get())) {
 						com.apocalypse.caerulaarbor.utils.WorldUtils.summonRandomSeaborn(world, rate, (double) x + 0.5, (double) y + 1.5, (double) z + 0.5);
 						{
 							int _value = 0;
