+++ src/main/java/com/apocalypse/caerulaarbor/block/SwarmcallerDollBlock.java
@@ -27,7 +27,7 @@ import java.util.Collections;
 import java.util.List;
 
 public class SwarmcallerDollBlock extends BaseEntityBlock implements SimpleWaterloggedBlock, EntityBlock {
-	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, (int) 1);
+	public static final IntegerProperty ANIMATION = IntegerProperty.create("animation", 0, 1);
 	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 
 	public SwarmcallerDollBlock() {
