+++ src/main/java/com/apocalypse/caerulaarbor/block/TrailLogBlock.java
@@ -1,7 +1,6 @@
 package com.apocalypse.caerulaarbor.block;
 
 import com.apocalypse.caerulaarbor.CaerulaArborMod;
-
 import com.apocalypse.caerulaarbor.init.CaerulaArborModBlocks;
 import com.apocalypse.caerulaarbor.procedures.PokeSlightlyProcedure;
 import net.minecraft.core.BlockPos;
@@ -93,7 +92,7 @@ public class TrailLogBlock extends Block {
         double longev = 0;
         if ((blockstate.getBlock().getStateDefinition().getProperty("grow_age") instanceof IntegerProperty _getip1 ? blockstate.getValue(_getip1) : -1) < 64) {
             {
-                int _value = (int) ((blockstate.getBlock().getStateDefinition().getProperty("grow_age") instanceof IntegerProperty _getip3 ? blockstate.getValue(_getip3) : -1) + 1);
+                int _value = (blockstate.getBlock().getStateDefinition().getProperty("grow_age") instanceof IntegerProperty _getip3 ? blockstate.getValue(_getip3) : -1) + 1;
                 BlockPos _pos = BlockPos.containing(x, y, z);
                 BlockState _bs = ((LevelAccessor) world).getBlockState(_pos);
                 if (_bs.getBlock().getStateDefinition().getProperty("grow_age") instanceof IntegerProperty _integerProp && _integerProp.getPossibleValues().contains(_value))
