+++ src/main/java/com/apocalypse/caerulaarbor/client/particle/IsharmlaCurseParticleParticle.java
@@ -1,16 +1,11 @@
 
 package com.apocalypse.caerulaarbor.client.particle;
 
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.api.distmarker.Dist;
-
-import net.minecraft.core.particles.SimpleParticleType;
-import net.minecraft.client.particle.TextureSheetParticle;
-import net.minecraft.client.particle.SpriteSet;
-import net.minecraft.client.particle.ParticleRenderType;
-import net.minecraft.client.particle.ParticleProvider;
-import net.minecraft.client.particle.Particle;
 import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.particle.*;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 
 @OnlyIn(Dist.CLIENT)
 public class IsharmlaCurseParticleParticle extends TextureSheetParticle {
@@ -37,7 +32,7 @@ public class IsharmlaCurseParticleParticle extends TextureSheetParticle {
 		this.spriteSet = spriteSet;
 		this.setSize(0.2f, 0.2f);
 		this.quadSize *= 1.5f;
-		this.lifetime = (int) Math.max(1, 35 + (this.random.nextInt(20) - 10));
+		this.lifetime = Math.max(1, 35 + (this.random.nextInt(20) - 10));
 		this.gravity = 0.35f;
 		this.hasPhysics = true;
 		this.xd = vx * 0.05;
